Catalog (#13)
* feat: create login viewmodel and fragment

* style: add corner radius, text style, and button style

* feat: add register fragment

* feat: login, register functionality

* fix merge conflict

* feat: major catalog

* style: add cancel button for search

* Revert "style: add cancel button for search"

This reverts commit 10d4b2feeaf8b547641956fb4e9fe02fb3e5c980.

* feat: add mutable live data for major catalog

* style: scroll back to top

* feat: observe session in main activity

* style: change height of navbar

* style: change searchbar style

* stye: list catalog

* style: add xml for major detail page

* feat: catalog detail page

+++ app/src/main/java/com/bangkit/nest/data/remote/retrofit/ApiService.kt
@@ -28,7 +28,7 @@ interface ApiService {
 
     @GET("major/detail/{id}")
     suspend fun getMajorDetail (
-        @Path("id") id: Long
+        @Path("id") id: Int
     ): DetailMajorResponse
